refactor: 디렉토리 구조 변경
Desc
- features 디렉토리를 app 디렉토리 내부에서 외부로 이동
- 공통 article 컴포넌트를 위치시킬 components 디렉토리 생성
