--- src/app/(features)/filter/style/filter.css
@@ -1,7 +1,0 @@
-@import 'tailwindcss';
-
-@layer components {
-  .filter-container {
-    @apply w-[100%] h-[60px] border-b-1 border-gray-300 sticky top-0;
-  }
-}
