+++ src/app/index.css
@@ -7,6 +7,7 @@
 
   body {
     height: 100%;
+    background-color: #f0f1f4;
   }
 }
