+++ src/app/scrap/scrap.css
@@ -1,0 +1,7 @@
+@import 'tailwindcss';
+
+@layer components {
+  .scrap-container {
+    height: 100%;
+  }
+}
