+++ src/components/article/index.tsx
@@ -1,0 +1,3 @@
+export const Article = () => {
+  return null;
+};
