+++ src/features/filter/style/filter.css
@@ -1,0 +1,7 @@
+@import 'tailwindcss';
+
+@layer components {
+  .filter-container {
+    @apply w-[100%] h-[60px] border-b-1 bg-white border-gray-300 sticky top-0;
+  }
+}
